fix: Suspense-compliant ServiceRegisterPage, framer-motion type fixes, MCP 구조 개선
